firmware: moved MPR121 sources into src/MPR121 subdir, removed some redundant includes
